Consolidate user-related APIs into UserApi
This commit moves the functionality from `AdminApi`, `PasswordApi`, and `RegistrationApi` into `UserApi`.

The following files and their corresponding tests have been deleted:
- `AdminApi.kt`
- `PasswordApi.kt`
- `RegistrationApi.kt`

The models previously located in `com.saintpatrck.mealie.client.api.admin.model`, `com.saintpatrck.mealie.client.api.password.model`, and `com.saintpatrck.mealie.client.api.registration.model` have been moved to `com.saintpatrck.mealie.client.api.user.model`.

`MealieClient` has been updated to reflect these changes, removing references to the deleted APIs.

`UserApi` now includes the following methods:
- `registerUser`
- `updateUserImage`
- `getUser`
- `deleteUser`
- `getAllUsers`
- `createUser`
- `forgotPassword`
- `resetPassword`

Tests in `UserApiTest.kt` have been updated to cover the newly added functionalities.
