+++ client/src/commonMain/kotlin/com/saintpatrck/mealie/client/api/user/model/RegisterUserResponseJson.kt
@@ -1,4 +1,4 @@
-package com.saintpatrck.mealie.client.api.registration.model
+package com.saintpatrck.mealie.client.api.user.model
 
 import com.saintpatrck.mealie.client.api.model.MealieToken
 import kotlinx.serialization.SerialName
