+++ client/src/commonMain/kotlin/com/saintpatrck/mealie/client/api/user/model/UpdateUserRequestJson.kt
@@ -1,6 +1,5 @@
 package com.saintpatrck.mealie.client.api.user.model
 
-import com.saintpatrck.mealie.client.api.registration.model.MealieAuthMethod
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
